moved search bar
