+++ src/components/monsters/MonsterPage.css
@@ -1,0 +1,28 @@
+.MonsterStatBlockRegion {
+    padding: 1%;
+}
+
+.remove-button {
+    text-align: end;
+    vertical-align: top;
+    position: absolute;
+    width: 20px;
+    height: 20px;
+    top: 0;
+    right: 0;
+    cursor: pointer;
+}
+
+.remove-button::after {
+    content: "";
+    position: absolute;
+    right: 3px;
+    top: 5px;
+    width: 10px;
+    height: 5px;
+    border-top: 2px solid rgba(0, 0, 0, 0.4);
+}
+
+.search-bar {
+    z-index: 10;
+}
