+++ src/components/monsters/MonsterSearchBar.css
@@ -26,7 +26,7 @@
 .react-autosuggest__suggestions-container--open {
     display: block;
     width: 100%;
-    z-index: 100;
+    z-index: 10;
     position: absolute;
     background-clip: padding-box;
     background-color: #FFFFFF;
